chore: fix permissions in dirty way for contacts vs cases
